refactor(backend): extract identity scoring primitives (719 -> 458 lines)
